reorganize sp1 folders
