+++ examples/clob/programs/Cargo.toml
@@ -7,4 +7,20 @@ edition = "2021"
 path = "src/lib.rs"
 
 [dependencies]
-clob-sp1 = { path = "sp1" }
+sp1-utils = { path = "../../../crates/sp1-utils" }
+once_cell = { workspace = true }
+
+[dev-dependencies]
+clob-core = { workspace = true }
+borsh = { workspace = true }
+alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
+ivm-zkvm = { workspace = true }
+ivm-abi = { workspace = true }
+clob-test-utils = { workspace = true }
+
+[build-dependencies]
+sp1-helper = { workspace = true }
+sp1-build = { workspace = true }
+sp1-sdk = { workspace = true }
+bincode = { workspace = true } 
+sp1-utils = { path = "../../../crates/sp1-utils" }
