+++ examples/clob/programs/program/Cargo.toml
@@ -9,9 +9,9 @@ edition = "2021"
 sp1-zkvm = { git = "https://github.com/succinctlabs/sp1.git", rev = "8a62caa0b8893aabe8e9477d4568c20387ae4896", features = ["libm", "lib"], default-features = false }
 alloy = { version = "0.2", features = ["sol-types"], default-features = false }
 borsh = { version = "1", default-features = false }
-clob-core = { path = "../../../core" }
-ivm-abi = { path = "../../../../../crates/sdk/abi" }
+clob-core = { path = "../../core" }
+ivm-abi = { path = "../../../../crates/sdk/abi" }
 
 [build-dependencies]
 sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", rev = "8a62caa0b8893aabe8e9477d4568c20387ae4896", default-features = false }
-sp1-build = "3.2.1" 
+sp1-build = "3.2.1"
