--- examples/clob/programs/sp1/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "clob-sp1"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-sp1-utils = { path = "../../../../crates/sp1-utils" }
-once_cell = { workspace = true }
-
-[dev-dependencies]
-clob-core = { workspace = true }
-borsh = { workspace = true }
-alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
-ivm-zkvm = { workspace = true }
-ivm-abi = { workspace = true }
-clob-test-utils = { workspace = true }
-
-[build-dependencies]
-sp1-helper = { workspace = true }
-sp1-build = { workspace = true }
-sp1-sdk = { workspace = true }
-bincode = { workspace = true } 
-sp1-utils = { path = "../../../../crates/sp1-utils" }
