+++ examples/matching-game/programs/Cargo.toml
@@ -7,4 +7,19 @@ edition = "2021"
 path = "src/lib.rs"
 
 [dependencies]
-matching-game-sp1 = { path = "sp1" }
+once_cell = { workspace = true }
+sp1-utils = { path = "../../../crates/sp1-utils" }
+ivm-zkvm = { workspace = true }
+
+[dev-dependencies]
+matching-game-core = { workspace = true }
+bincode = { workspace = true }
+alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
+ivm-abi = { workspace = true }
+kairos-trie = { workspace = true }
+
+[build-dependencies]
+sp1-utils = { path = "../../../crates/sp1-utils" }
+sp1-build = { workspace = true }
+sp1-sdk = { workspace = true }
+bincode = { workspace = true }
