+++ examples/matching-game/programs/program/Cargo.toml
@@ -9,8 +9,8 @@ edition = "2021"
 sp1-zkvm = { git = "https://github.com/succinctlabs/sp1.git", rev = "8a62caa0b8893aabe8e9477d4568c20387ae4896", features = ["libm", "lib"], default-features = false }
 alloy = { version = "0.2", features = ["sol-types"], default-features = false }
 bincode = { version = "1.3", default-features = false }
-matching-game-core = { path = "../../../core" }
-ivm-abi = { path = "../../../../../crates/sdk/abi" }
+matching-game-core = { path = "../../core" }
+ivm-abi = { path = "../../../../crates/sdk/abi" }
 sha2 = { version = "0.10", default-features = false }
 kairos-trie = { git = "https://github.com/cspr-rad/kairos-trie.git", features = ["serde"] }
