--- examples/matching-game/programs/sp1/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "matching-game-sp1"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-once_cell = { workspace = true }
-sp1-utils = { path = "../../../../crates/sp1-utils" }
-ivm-zkvm = { workspace = true }
-
-[dev-dependencies]
-matching-game-core = { workspace = true }
-bincode = { workspace = true }
-alloy = { workspace = true, features = ["full", "contract", "transport-http", "transport-ws", "network", "providers", "pubsub", "signer-local", "signer-keystore", "rpc-types-debug", "rpc-types-trace", "json-rpc", "rpc-client", "rlp", "sol-types"] }
-ivm-abi = { workspace = true }
-kairos-trie = { workspace = true }
-
-[build-dependencies]
-sp1-utils = { path = "../../../../crates/sp1-utils" }
-sp1-build = { workspace = true }
-sp1-sdk = { workspace = true }
-bincode = { workspace = true }
