+++ programs/intensity-test/build.rs
@@ -1,0 +1,5 @@
+use sp1_utils::build_sp1_program;
+
+fn main() {
+    build_sp1_program("intensity-test-sp1-guest", "program", "programs/intensity-test/program/elf");
+}
