--- programs/intensity-test/sp1/build.rs
@@ -1,9 +1,0 @@
-use sp1_utils::build_sp1_program;
-
-fn main() {
-    build_sp1_program(
-        "intensity-test-sp1-guest",
-        "program",
-        "programs/intensity-test/sp1/program/elf",
-    );
-}
