+++ programs/mock-consumer/program/src/main.rs
