--- programs/mock-consumer/sp1/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "mock-consumer-sp1"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-once_cell = { workspace = true, features = ["std"] }
-sp1-utils = { path = "../../../crates/sp1-utils" }
-
-[build-dependencies]
-sp1-helper = { workspace = true }
-sp1-build = { workspace = true }
-sp1-sdk = { workspace = true }
-sp1-utils = { path = "../../../crates/sp1-utils" }
