--- programs/mock-consumer/sp1/build.rs
@@ -1,9 +1,0 @@
-use sp1_utils::build_sp1_program;
-
-fn main() {
-    build_sp1_program(
-        "mock-consumer-sp1-guest",
-        "program",
-        "programs/mock-consumer/sp1/program/elf/",
-    );
-}
